Patch to fix crc32 (-softcrc) under 64-bit linux

+++ src/misc/crc32.c
@@ -17,14 +17,14 @@ $Id: crc32.c,v 1.2 2006/01/02 18:24:04 rasc Exp $
 
 
 
-#include <sys/types.h>
+#include <stdint.h>
 #include "crc32.h"
 
 
 
 // CRC32 lookup table for polynomial 0x04c11db7
 
-static u_long crc_table[256] = {
+static uint32_t crc_table[256] = {
 	0x00000000, 0x04c11db7, 0x09823b6e, 0x0d4326d9, 0x130476dc, 0x17c56b6b,
 	0x1a864db2, 0x1e475005, 0x2608edb8, 0x22c9f00f, 0x2f8ad6d6, 0x2b4bcb61,
 	0x350c9b64, 0x31cd86d3, 0x3c8ea00a, 0x384fbdbd, 0x4c11db70, 0x48d0c6c7,
@@ -69,10 +69,10 @@ static u_long crc_table[256] = {
 	0x933eb0bb, 0x97ffad0c, 0xafb010b1, 0xab710d06, 0xa6322bdf, 0xa2f33668,
 	0xbcb4666d, 0xb8757bda, 0xb5365d03, 0xb1f740b4};
 
-u_long crc32 (char *data, int len)
+uint32_t crc32 (char *data, int len)
 {
 	register int i;
-	u_long crc = 0xffffffff;
+	uint32_t crc = 0xffffffff;
 
 	for (i=0; i<len; i++)
 		crc = (crc << 8) ^ crc_table[((crc >> 24) ^ *data++) & 0xff];
